Rename package with 'properties'
