+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/diagnostic/BLangDiagnostic.java
@@ -21,7 +21,7 @@ import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticCode;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
 import io.ballerina.tools.diagnostics.Location;
-import io.ballerina.tools.diagnostics.arguments.DiagnosticProperty;
+import io.ballerina.tools.diagnostics.properties.DiagnosticProperty;
 
 import java.util.Collections;
 import java.util.List;
