+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/diagnostic/BLangDiagnosticLog.java
@@ -26,15 +26,15 @@ import io.ballerina.tools.diagnostics.DiagnosticCode;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 import io.ballerina.tools.diagnostics.Location;
-import io.ballerina.tools.diagnostics.arguments.DiagnosticProperty;
+import io.ballerina.tools.diagnostics.properties.DiagnosticProperty;
 import org.ballerinalang.model.elements.PackageID;
 import org.ballerinalang.util.diagnostic.DiagnosticLog;
 import org.wso2.ballerinalang.compiler.PackageCache;
-import org.wso2.ballerinalang.compiler.diagnostic.arguments.BCollectionProperty;
-import org.wso2.ballerinalang.compiler.diagnostic.arguments.BNumericProperty;
-import org.wso2.ballerinalang.compiler.diagnostic.arguments.BStringProperty;
-import org.wso2.ballerinalang.compiler.diagnostic.arguments.BSymbolicProperty;
-import org.wso2.ballerinalang.compiler.diagnostic.arguments.NonCatProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BCollectionProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BNumericProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BStringProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BSymbolicProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.NonCatProperty;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.BSymbol;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BType;
 import org.wso2.ballerinalang.compiler.tree.BLangPackage;
