+++ compiler/ballerina-tools-api/src/main/java/io/ballerina/tools/diagnostics/properties/CollectionProperty.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.tools.diagnostics.arguments;
+package io.ballerina.tools.diagnostics.properties;
 
 /**
  * Represents collection of properties passed when diagnostic logging.
