+++ compiler/ballerina-tools-api/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module io.ballerina.tools.api {
     exports io.ballerina.tools.diagnostics;
-    exports io.ballerina.tools.diagnostics.arguments;
+    exports io.ballerina.tools.diagnostics.properties;
     exports io.ballerina.tools.text;
 }
