+++ misc/toml-parser/src/main/java/io/ballerina/toml/internal/diagnostics/SyntaxDiagnostic.java
@@ -22,7 +22,7 @@ import io.ballerina.toml.syntax.tree.NodeLocation;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticCode;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
-import io.ballerina.tools.diagnostics.arguments.DiagnosticProperty;
+import io.ballerina.tools.diagnostics.properties.DiagnosticProperty;
 
 import java.util.List;
