+++ misc/toml-parser/src/main/java/io/ballerina/toml/semantic/diagnostics/TomlDiagnostic.java
@@ -21,7 +21,7 @@ package io.ballerina.toml.semantic.diagnostics;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
 import io.ballerina.tools.diagnostics.Location;
-import io.ballerina.tools.diagnostics.arguments.DiagnosticProperty;
+import io.ballerina.tools.diagnostics.properties.DiagnosticProperty;
 
 import java.util.List;
